[8.0][FIX] l10n_it_fatturapa_pec, l10n_it_sdi_channel: better fetchmail handling
PREMISE: When an electronic invoice is sent via PEC using send_via_pec
method, a fetchmail check is performed each time. This check is useless
because the fetchmail is set up once when module is installed.
Thus, we added pec_fetchmail_server_id field into sdi.channel model and
set it required when channel type is PEC. In this way, after installation
the user must set up ingoing and outgoing email server in order to enable
correctly the sdi channel.

  * [ADD] added pec_fetchmail_server_id in sdi PEC channel,
  in order to identify the fetchmail server used to retrieve
  SDI response messages.

  * [IMP] improved sdi.channel form view, setting PEC field required
  when the channel type is PEC

  * [FIX] removed _check_fetchmail check during send_via_pec method
  (see premise)